backend code completed for booking

+++ src/main/java/com/example/HotelBookingSystem/dto/RoomsResponseDTO.java
@@ -3,9 +3,9 @@ package com.example.HotelBookingSystem.dto;
 public record RoomsResponseDTO(int id,
                                String roomType,
                                int rate,
-                               Integer gstPercent,    // optional Integer supports null
-                               Integer acRate,        // optional
-                               Integer extraBedFee,   // optional
+                               Double gstPercent,    // optional Integer supports null
+                               Double acRate,        // optional
+                               Double extraBedFee,   // optional
                                String remarks,        // optional
                                int noOfRooms) {
 }
